backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["SilasBach <79838393+SilasBach@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
datetime = "^5.4"
fastapi = "^0.105.0"
uvicorn = "^0.24.0.post1"
pinecone-client = "^2.2.4"
langchain = "^0.0.351"
pypdf = "^3.17.3"
sentence-transformers = "^2.2.2"
replicate = "^0.22.0"
pymongo = "^4.6.1"
motor = "^3.3.2"
pydantic = {extras = ["email"], version = "^2.5.3"}
pyjwt = {extras = ["encode"], version = "^2.8.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
